익셉션 처리

+++ src/main/java/common/util/file/CommonsFileUtil.java
@@ -1,5 +1,5 @@
 /**
- * 
+ *
  */
 package common.util.file;
 
@@ -18,6 +18,7 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.filefilter.DirectoryFileFilter;
 import org.apache.commons.io.filefilter.NotFileFilter;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -32,14 +33,14 @@ import org.slf4j.LoggerFactory;
  * </pre>
  */
 public class CommonsFileUtil {
-	
+
 	private static final Logger logger = LoggerFactory.getLogger(CommonsFileUtil.class);
-	
+
 	/**
 	 * 폴더 구분자
 	 */
 	public static final String FOLDER_SEPARATOR = "/";
-	
+
 	/**
 	 * 확장자 구분자
 	 */
@@ -48,7 +49,7 @@ public class CommonsFileUtil {
 	private CommonsFileUtil() {
 		super();
 	}
-	
+
 	/**
 	 * 파일의 존재여부 확인
 	 * @param filePath
@@ -56,25 +57,37 @@ public class CommonsFileUtil {
 	 * @since 1.7
 	 */
 	public static boolean isExistsFile(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		Path path = Paths.get(filePath);
         return path.toFile().exists();
     }
-	
+
 	/**
 	 * 해당 경로에서 파일명 추출
 	 * @param filePath
 	 * @return
 	 */
 	public static String getFilename(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		return FilenameUtils.getBaseName(filePath);
 	}
-	
+
 	/**
 	 * 파일 확장자 구하기
 	 * @param fileName
 	 * @return
 	 */
 	public static String getFileExtension(String fileName) {
+		if ( StringUtils.isBlank(fileName) ) {
+			throw new NullPointerException("fileName is null");
+		}
+
 		return FilenameUtils.getExtension(fileName);
 	}
 
@@ -84,10 +97,14 @@ public class CommonsFileUtil {
 	 * @return
 	 */
 	public static long getFileSize(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		File file = FileUtils.getFile(filePath);
 		return FileUtils.sizeOf(file);
 	}
-	
+
 	/**
 	 * <pre>
 	 * 파일 용량 구하기
@@ -97,36 +114,52 @@ public class CommonsFileUtil {
 	 * @return
 	 */
 	public static String readableFileSize(long fileSize) {
+		if ( fileSize < 0 ) {
+			throw new IllegalArgumentException("fileSize is negative");
+		}
+
 	    return FileUtils.byteCountToDisplaySize(fileSize);
 	}
-	
+
 	/**
 	 * 파일의 수정한 날짜 구하기
 	 * @param filename
 	 * @return
 	 */
 	public static String lastModified(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		File file = FileUtils.getFile(filePath);
 		Date date = new Date(file.lastModified());
 		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
 		return sdf.format(date);
 	}
-	
+
 	/**
 	 * 텍스트 내용을 행당 경로에 파일로 생성
 	 * @param filePath
 	 * @param text
 	 */
 	public static void writeFile(String filePath, String text) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
+		if ( StringUtils.isBlank(text) ) {
+			throw new NullPointerException("text is null");
+		}
+
 		try {
 			File file = FileUtils.getFile(filePath);
 			FileUtils.writeStringToFile(file, text, Charset.defaultCharset());
-			
+
 		} catch (IOException e) {
 			logger.error("", e);
-		}	
+		}
 	}
-	
+
 	/**
 	 * 텍스트 내용을 행당 경로에 파일로 생성
 	 * @param filePath
@@ -134,34 +167,50 @@ public class CommonsFileUtil {
 	 * @param encoding
 	 */
 	public static void writeFile(String filePath, String text, String encoding) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
+		if ( StringUtils.isBlank(text) ) {
+			throw new NullPointerException("text is null");
+		}
+
+		if ( StringUtils.isBlank(encoding) ) {
+			throw new NullPointerException("encoding is null");
+		}
+
 		try {
 			File file = FileUtils.getFile(filePath);
 			FileUtils.writeStringToFile(file, text, encoding);
-			
+
 		} catch (IOException e) {
 			logger.error("", e);
-		}		
+		}
 	}
-	
+
 	/**
 	 * 파일을 텍스트로 읽음
 	 * @param filePath
 	 * @return
 	 */
 	public static String readFile(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		String content = "";
-		
+
 		try {
 			File file = FileUtils.getFile(filePath);
 			content = FileUtils.readFileToString(file, Charset.defaultCharset());
-			
+
 		} catch (IOException e) {
 			logger.error("", e);
 		}
-		
+
 		return content;
 	}
-	
+
 	/**
 	 * 파일을 텍스트로 읽음
 	 * @param filePath
@@ -169,116 +218,152 @@ public class CommonsFileUtil {
 	 * @return
 	 */
 	public static String readFile(String filePath, String encoding) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
+		if ( StringUtils.isBlank(encoding) ) {
+			throw new NullPointerException("encoding is null");
+		}
+
 		String content = "";
-		
+
 		try {
 			File file = FileUtils.getFile(filePath);
 			content = FileUtils.readFileToString(file, encoding);
-			
+
 		} catch (IOException e) {
 			logger.error("", e);
 		}
-		
+
 		return content;
 	}
-	
+
 	/**
 	 * 파일 삭제
 	 * @param filePath
 	 */
 	public static boolean deleteFile(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		File file = FileUtils.getFile(filePath);
 		return FileUtils.deleteQuietly(file);
 	}
-	
+
 	/**
 	 * 파일 복사
 	 * @param srcFilePath
 	 * @param destFilePath
 	 */
 	public static void copyFile(String srcFilePath, String destFilePath) {
+		if ( StringUtils.isBlank(srcFilePath) ) {
+			throw new NullPointerException("srcFilePath is null");
+		}
+
+		if ( StringUtils.isBlank(destFilePath) ) {
+			throw new NullPointerException("destFilePath is null");
+		}
+
 		File srcFile = FileUtils.getFile(srcFilePath);
 		File destFile = FileUtils.getFile(destFilePath);
-		
+
 		try {
 			FileUtils.copyFile(srcFile, destFile);
-			
+
 		} catch (IOException e) {
 			logger.error("", e);
 		}
 	}
-	
+
 	/**
 	 * 해당 경로의 모든 파일 및 디렉토리를 반환
 	 * @param path
 	 * @return
 	 */
 	public static List<String> getAllFileList(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		List<String> listFiles = new ArrayList<>();
 		File file = FileUtils.getFile(filePath);
-		
+
 		File[] files = file.listFiles();
 		for (File f : files) {
 			listFiles.add(f.getName());
 		}
-		
+
 		return listFiles;
 	}
-	
+
 	/**
 	 * 해당 경로의 파일 반환
 	 * @param path
 	 * @return
 	 */
 	public static List<String> getFileList(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		List<String> listFiles = new ArrayList<>();
 		File file = FileUtils.getFile(filePath);
-		
-		File[] files = file.listFiles( 
+
+		File[] files = file.listFiles(
 				(FilenameFilter) new NotFileFilter(DirectoryFileFilter.DIRECTORY) );
-		
+
 		for (File f : files) {
 			listFiles.add(f.getName());
 		}
-		
+
 		return listFiles;
 	}
-	
+
 	/**
 	 * 해당 경로의 디렉토리 반환
 	 * @param path
 	 * @return
 	 */
 	public static List<String> getDirectoryList(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		List<String> listFiles = new ArrayList<>();
 		File file = FileUtils.getFile(filePath);
-		
-		File[] files = file.listFiles( 
+
+		File[] files = file.listFiles(
 				(FilenameFilter) DirectoryFileFilter.DIRECTORY );
-		
+
 		for (File f : files) {
 			listFiles.add(f.getName());
 		}
-		
+
 		return listFiles;
 	}
-	
+
 	/**
 	 * 파일을 byte[]로 변환
 	 * @param filePath
 	 * @return
 	 */
 	public static byte[] convertFileToBytes(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		byte[] bData = null;
-		
+
 		File file = FileUtils.getFile(filePath);
-		
+
 		try {
 			bData = FileUtils.readFileToByteArray(file);
 		} catch (IOException e) {
 			logger.error("", e);
 		}
-		
+
 		return bData;
 	}
