+++ src/main/java/common/util/file/NioFileUtil.java
@@ -16,13 +16,14 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * @author 김대광
  * @Description : 1.7 기반
- * 
+ *
  *              <pre>
  * -----------------------------------
  * 개정이력
@@ -48,22 +49,30 @@ public class NioFileUtil {
 
 	/**
 	 * 파일의 존재여부 확인
-	 * 
+	 *
 	 * @param filePath
 	 * @return
 	 */
 	public static boolean isExistsFile(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		Path path = Paths.get(filePath);
 		return path.toFile().exists();
 	}
 
 	/**
 	 * 해당 경로에서 파일명 추출
-	 * 
+	 *
 	 * @param filePath
 	 * @return
 	 */
 	public static String getFilename(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		if (filePath == null) {
 			return null;
 		}
@@ -73,11 +82,15 @@ public class NioFileUtil {
 
 	/**
 	 * 파일 확장자 구하기
-	 * 
+	 *
 	 * @param fileName
 	 * @return
 	 */
 	public static String getFileExtension(String fileName) {
+		if ( StringUtils.isBlank(fileName) ) {
+			throw new NullPointerException("fileName is null");
+		}
+
 		if (fileName.lastIndexOf(EXTENSION_SEPARATOR) == -1) {
 			return null;
 		}
@@ -87,11 +100,15 @@ public class NioFileUtil {
 
 	/**
 	 * 파일 용량 구하기
-	 * 
+	 *
 	 * @param filePath
 	 * @return
 	 */
 	public static long getFileSize(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		Path path = Paths.get(filePath);
 
 		try {
@@ -107,11 +124,15 @@ public class NioFileUtil {
 	 * 파일 용량 구하기
 	 *   - B, KB, MB, GB, TB
 	 * </pre>
-	 * 
+	 *
 	 * @param fileSize
 	 * @return
 	 */
 	public static String readableFileSize(long fileSize) {
+		if ( fileSize < 0 ) {
+			throw new IllegalArgumentException("fileSize is negative");
+		}
+
 		if (fileSize <= 0)
 			return "0";
 		String[] units = { "B", "KB", "MB", "GB", "TB" };
@@ -122,11 +143,15 @@ public class NioFileUtil {
 
 	/**
 	 * 파일의 수정한 날짜 구하기
-	 * 
+	 *
 	 * @param filePath
 	 * @return
 	 */
 	public static String lastModified(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		Path path = Paths.get(filePath);
 
 		FileTime time = null;
@@ -144,11 +169,19 @@ public class NioFileUtil {
 
 	/**
 	 * 텍스트 내용을 행당 경로에 파일로 생성
-	 * 
+	 *
 	 * @param filePath
 	 * @param text
 	 */
 	public static void writeFile(String filePath, String text) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
+		if ( StringUtils.isBlank(text) ) {
+			throw new NullPointerException("text is null");
+		}
+
 		Path path = Paths.get(filePath);
 
 		try (InputStream is = new ByteArrayInputStream(text.getBytes())) {
@@ -162,12 +195,24 @@ public class NioFileUtil {
 
 	/**
 	 * 텍스트 내용을 행당 경로에 파일로 생성
-	 * 
+	 *
 	 * @param filePath
 	 * @param text
 	 * @param encoding
 	 */
 	public static void writeFile(String filePath, String text, String encoding) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
+		if ( StringUtils.isBlank(text) ) {
+			throw new NullPointerException("text is null");
+		}
+
+		if ( StringUtils.isBlank(encoding) ) {
+			throw new NullPointerException("encoding is null");
+		}
+
 		Path targetFile = Paths.get(filePath);
 
 		try (InputStream is = new ByteArrayInputStream(text.getBytes(encoding))) {
@@ -181,11 +226,19 @@ public class NioFileUtil {
 
 	/**
 	 * bytes를 해당 경로에 파일로 생성
-	 * 
+	 *
 	 * @param filePath
 	 * @param bData
 	 */
 	public static void writeFile(String filePath, byte[] bData) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
+		if ( bData == null || bData.length == 0 ) {
+			throw new NullPointerException("bData is null");
+		}
+
 		Path path = Paths.get(filePath);
 		try {
 			Files.write(path, bData);
@@ -197,11 +250,15 @@ public class NioFileUtil {
 
 	/**
 	 * 파일을 텍스트로 읽음
-	 * 
+	 *
 	 * @param filePath
 	 * @return
 	 */
 	public static String readFile(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		String content = "";
 		byte[] bData = null;
 
@@ -220,12 +277,20 @@ public class NioFileUtil {
 
 	/**
 	 * 파일을 텍스트로 읽음
-	 * 
+	 *
 	 * @param filePath
 	 * @param encoding
 	 * @return
 	 */
 	public static String readFile(String filePath, String encoding) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
+		if ( StringUtils.isBlank(encoding) ) {
+			throw new NullPointerException("encoding is null");
+		}
+
 		String content = "";
 		byte[] bData = null;
 
@@ -248,30 +313,34 @@ public class NioFileUtil {
 
 	/**
 	 * 파일 삭제
-	 * 
+	 *
 	 * @param filePath
 	 */
 	public static boolean deleteFile(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		Path path = Paths.get(filePath);
-		
+
 		try {
 			if ( path.toFile().isDirectory() ) {
 				File[] files = path.toFile().listFiles();
-				
+
 				for (File f : files) {
 					deleteFile(f.getPath());
 					logger.debug("파일이 삭제되었습니다.");
 				}
-				
+
 				Files.delete(path);
 				logger.debug("폴더가 삭제되었습니다.");
 			} else {
 				Files.delete(path);
-				logger.debug("파일이 삭제되었습니다.");				
+				logger.debug("파일이 삭제되었습니다.");
 			}
-			
+
 			return true;
-			
+
 		} catch (IOException e) {
 			return false;
 		}
@@ -279,11 +348,19 @@ public class NioFileUtil {
 
 	/**
 	 * 파일 복사
-	 * 
+	 *
 	 * @param srcFilePath
 	 * @param destFilePath
 	 */
 	public static void copyFile(String srcFilePath, String destFilePath) {
+		if ( StringUtils.isBlank(srcFilePath) ) {
+			throw new NullPointerException("srcFilePath is null");
+		}
+
+		if ( StringUtils.isBlank(destFilePath) ) {
+			throw new NullPointerException("destFilePath is null");
+		}
+
 		Path srcFile = Paths.get(srcFilePath);
 		Path destFile = Paths.get(destFilePath);
 
@@ -297,11 +374,15 @@ public class NioFileUtil {
 
 	/**
 	 * 해당 경로의 모든 파일 및 디렉토리를 반환
-	 * 
+	 *
 	 * @param filePath
 	 * @return
 	 */
 	public static List<String> getAllFileList(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		List<String> listFiles = new ArrayList<>();
 		Path path = Paths.get(filePath);
 
@@ -322,11 +403,15 @@ public class NioFileUtil {
 
 	/**
 	 * 해당 경로의 파일 반환
-	 * 
+	 *
 	 * @param filePath
 	 * @return
 	 */
 	public static List<String> getFileList(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		List<String> listFiles = new ArrayList<>();
 		Path path = Paths.get(filePath);
 
@@ -351,11 +436,15 @@ public class NioFileUtil {
 
 	/**
 	 * 해당 경로의 디렉토리 반환
-	 * 
+	 *
 	 * @param filePath
 	 * @return
 	 */
 	public static List<String> getDirectoryList(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		List<String> listFiles = new ArrayList<>();
 		Path path = Paths.get(filePath);
 
@@ -378,11 +467,15 @@ public class NioFileUtil {
 
 	/**
 	 * 파일을 byte[]로 변환
-	 * 
+	 *
 	 * @param filePath
 	 * @return
 	 */
 	public static byte[] convertFileToBytes(String filePath) {
+		if ( StringUtils.isBlank(filePath) ) {
+			throw new NullPointerException("filePath is null");
+		}
+
 		byte[] bData = null;
 
 		try {
